project/requirements.txt
appnope==0.1.0
backcall==0.1.0
click==7.1.1
decorator==4.4.2
dj-database-url==0.5.0
Django==2.2
django-countries==6.1.2
django-heroku==0.3.1
django-timezone-field==4.0
djangorestframework==3.11.0
gunicorn==20.0.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
nltk==3.5
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.4
ptyprocess==0.6.0
pyfiglet==0.8.post1
Pygments==2.6.1
PyInquirer==1.0.3
pytz==2019.3
regex==2020.2.20
six==1.14.0
sqlparse==0.3.1
textblob==0.15.3
tqdm==4.45.0
traitlets==4.3.3
wcwidth==0.1.9
whitenoise==5.0.1
